requirements.txt
openai-whisper
pyaudio
# `brew install portaudio` may also be necessary